added lifes

+++ src/com/codepenguins/rolling/io/Render.java
@@ -1,8 +1,10 @@
 package com.codepenguins.rolling.io;
 
 import com.codepenguins.rolling.model.GameObject;
+import com.codepenguins.rolling.model.UiObject;
 
 import static org.lwjgl.opengl.GL11.*;
+
 import org.lwjgl.LWJGLException;
 import org.lwjgl.opengl.Display;
 import org.lwjgl.opengl.DisplayMode;
@@ -52,13 +54,13 @@ public class Render {
 	public void drawObject(GameObject obj) {
 		Tex.getTexture(obj.getTextureId()).bind();
 		glBegin(GL_QUADS);
-			glTexCoord2f(((float)obj.getFrameId()) / obj.getFrameCount(), 0);
+			glTexCoord2f(obj.getFrameId() / (float) obj.getFrameCount(), 0);
 			glVertex3f(obj.getX(), obj.getY(), -1);
-			glTexCoord2f(((float)obj.getFrameId()) / obj.getFrameCount(), 1);
+			glTexCoord2f(obj.getFrameId() / (float) obj.getFrameCount(), 1);
 			glVertex3f(obj.getX(), obj.getY()+obj.getHeight(), -1);
-			glTexCoord2f((obj.getFrameId() + 1.0f) / obj.getFrameCount(), 1);
+			glTexCoord2f((obj.getFrameId() + 1) / (float) obj.getFrameCount(), 1);
 			glVertex3f(obj.getX()+obj.getWidth(), obj.getY()+obj.getHeight(), -1);
-			glTexCoord2f((obj.getFrameId() + 1.0f) / obj.getFrameCount(), 0);
+			glTexCoord2f((obj.getFrameId() + 1) / (float) obj.getFrameCount(), 0);
 			glVertex3f(obj.getX()+obj.getWidth(), obj.getY(), -1);
 		glEnd();
 	}
